remove InputReceiver
