+++ com.unity.renderstreaming/Editor/InputSystem/SimplePlayerInputEditor.cs
@@ -29,6 +29,9 @@ namespace Unity.RenderStreaming.InputSystem.Editor
         {
             EditorGUI.BeginChangeCheck();
 
+            EditorGUILayout.PropertyField(serializedObject.FindProperty("local"));
+            EditorGUILayout.PropertyField(serializedObject.FindProperty("label"));
+
             // Action config section.
             EditorGUI.BeginChangeCheck();
             EditorGUILayout.PropertyField(m_ActionsProperty);
