--- com.unity.renderstreaming/Runtime/Scripts/InputReceiver.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0f8d5d29a31dd084aaa8cb13851d7ef8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
